MAINT: fix tests after moving MatVecOperator
