+++ doc/source/getting_started/code/getting_started_convolution.py
@@ -2,6 +2,7 @@
 
 import odl
 import scipy
+import scipy.signal
 
 
 class Convolution(odl.Operator):
