+++ odl/test/space/ntuples_test.py
@@ -26,7 +26,7 @@ from builtins import range
 import numpy as np
 import pytest
 import operator
-import scipy as sp
+import scipy
 
 # ODL imports
 import odl
@@ -37,8 +37,7 @@ from odl.space.npy_ntuples import (
     NumpyFnConstWeighting, NumpyFnArrayWeighting, NumpyFnMatrixWeighting,
     NumpyFnNoWeighting, NumpyFnCustomInner, NumpyFnCustomNorm,
     NumpyFnCustomDist,
-    npy_weighted_inner, npy_weighted_norm, npy_weighted_dist,
-    MatVecOperator)
+    npy_weighted_inner, npy_weighted_norm, npy_weighted_dist)
 from odl.util.testutils import (almost_equal, all_almost_equal, all_equal,
                                 noise_array, noise_element,
                                 noise_elements, simple_fixture)
@@ -74,7 +73,7 @@ def _dense_matrix(fn):
 
 def _sparse_matrix(fn):
     """Create a sparse positive definite Hermitian matrix for `fn`."""
-    return sp.sparse.coo_matrix(_dense_matrix(fn))
+    return scipy.sparse.coo_matrix(_dense_matrix(fn))
 
 
 # Pytest fixtures
@@ -810,208 +809,6 @@ def test_conj(fn):
     assert all_equal(y, xarr.conj())
 
 
-# ---- MatVecOperator ---- #
-
-
-def test_matvec_init(fn):
-    # Square matrices, sparse and dense
-    sparse_mat = _sparse_matrix(fn)
-    dense_mat = _dense_matrix(fn)
-
-    MatVecOperator(sparse_mat, fn, fn)
-    MatVecOperator(dense_mat, fn, fn)
-
-    # Test defaults
-    op_float = MatVecOperator([[1.0, 2],
-                               [-1, 0.5]])
-
-    assert isinstance(op_float.domain, NumpyFn)
-    assert op_float.domain.is_rn
-    assert isinstance(op_float.range, NumpyFn)
-    assert op_float.domain.is_rn
-
-    op_complex = MatVecOperator([[1.0, 2 + 1j],
-                                 [-1 - 1j, 0.5]])
-
-    assert isinstance(op_complex.domain, NumpyFn)
-    assert op_complex.domain.is_cn
-    assert isinstance(op_complex.range, NumpyFn)
-    assert op_complex.domain.is_cn
-
-    op_int = MatVecOperator([[1, 2],
-                             [-1, 0]])
-
-    assert isinstance(op_int.domain, NumpyFn)
-    assert op_int.domain.dtype == int
-    assert isinstance(op_int.range, NumpyFn)
-    assert op_int.domain.dtype == int
-
-    # Rectangular
-    rect_mat = 2 * np.eye(2, 3)
-    r2 = odl.rn(2)
-    r3 = odl.rn(3)
-
-    MatVecOperator(rect_mat, r3, r2)
-
-    with pytest.raises(ValueError):
-        MatVecOperator(rect_mat, r2, r2)
-
-    with pytest.raises(ValueError):
-        MatVecOperator(rect_mat, r3, r3)
-
-    with pytest.raises(ValueError):
-        MatVecOperator(rect_mat, r2, r3)
-
-    # Rn to Cn okay
-    MatVecOperator(rect_mat, r3, odl.cn(2))
-
-    # Cn to Rn not okay (no safe cast)
-    with pytest.raises(TypeError):
-        MatVecOperator(rect_mat, odl.cn(3), r2)
-
-    # Complex matrix between real spaces not okay
-    rect_complex_mat = rect_mat + 1j
-    with pytest.raises(TypeError):
-        MatVecOperator(rect_complex_mat, r3, r2)
-
-    # Init with array-like structure (including numpy.matrix)
-    MatVecOperator(rect_mat.tolist(), r3, r2)
-    MatVecOperator(np.asmatrix(rect_mat), r3, r2)
-
-
-def test_matvec_simple_properties():
-    # Matrix - always ndarray in for dense input, scipy.sparse.spmatrix else
-    rect_mat = 2 * np.eye(2, 3)
-    r2 = odl.rn(2)
-    r3 = odl.rn(3)
-
-    op = MatVecOperator(rect_mat, r3, r2)
-    assert isinstance(op.matrix, np.ndarray)
-
-    op = MatVecOperator(np.asmatrix(rect_mat), r3, r2)
-    assert isinstance(op.matrix, np.ndarray)
-
-    op = MatVecOperator(rect_mat.tolist(), r3, r2)
-    assert isinstance(op.matrix, np.ndarray)
-    assert not op.matrix_issparse
-
-    sparse_mat = _sparse_matrix(odl.rn(5))
-    op = MatVecOperator(sparse_mat, odl.rn(5), odl.rn(5))
-    assert isinstance(op.matrix, sp.sparse.spmatrix)
-    assert op.matrix_issparse
-
-
-def test_matvec_adjoint(fn):
-    # Square cases
-    sparse_mat = _sparse_matrix(fn)
-    dense_mat = _dense_matrix(fn)
-
-    op_sparse = MatVecOperator(sparse_mat, fn, fn)
-    op_dense = MatVecOperator(dense_mat, fn, fn)
-
-    # Just test if it runs, nothing interesting to test here
-    op_sparse.adjoint
-    op_dense.adjoint
-
-    # Rectangular case
-    rect_mat = 2 * np.eye(2, 3)
-    r2, r3 = odl.rn(2), odl.rn(3)
-    c2 = odl.cn(2)
-
-    op = MatVecOperator(rect_mat, r3, r2)
-    op_adj = op.adjoint
-    assert op_adj.domain == op.range
-    assert op_adj.range == op.domain
-    assert np.array_equal(op_adj.matrix, op.matrix.conj().T)
-    assert np.array_equal(op_adj.adjoint.matrix, op.matrix)
-
-    # The operator Rn -> Cn has no adjoint
-    op_noadj = MatVecOperator(rect_mat, r3, c2)
-    with pytest.raises(NotImplementedError):
-        op_noadj.adjoint
-
-
-def test_matvec_inverse(fn):
-    # Sparse case
-    sparse_mat = _sparse_matrix(fn)
-    op_sparse = MatVecOperator(sparse_mat, fn, fn)
-
-    op_sparse_inv = op_sparse.inverse
-    assert op_sparse_inv.domain == op_sparse.range
-    assert op_sparse_inv.range == op_sparse.domain
-    assert all_almost_equal(op_sparse_inv.matrix,
-                            np.linalg.inv(op_sparse.matrix.todense()))
-    assert all_almost_equal(op_sparse_inv.inverse.matrix,
-                            op_sparse.matrix.todense())
-
-    # Test application
-    x = noise_element(fn)
-    assert all_almost_equal(x, op_sparse.inverse(op_sparse(x)))
-
-    # Dense case
-    dense_mat = _dense_matrix(fn)
-    op_dense = MatVecOperator(dense_mat, fn, fn)
-    op_dense_inv = op_dense.inverse
-    assert op_dense_inv.domain == op_dense.range
-    assert op_dense_inv.range == op_dense.domain
-    assert all_almost_equal(op_dense_inv.matrix,
-                            np.linalg.inv(op_dense.matrix))
-    assert all_almost_equal(op_dense_inv.inverse.matrix,
-                            op_dense.matrix)
-
-    # Test application
-    x = noise_element(fn)
-    assert all_almost_equal(x, op_dense.inverse(op_dense(x)))
-
-
-def test_matvec_call(fn):
-    # Square cases
-    sparse_mat = _sparse_matrix(fn)
-    dense_mat = _dense_matrix(fn)
-    xarr, x = noise_elements(fn)
-
-    op_sparse = MatVecOperator(sparse_mat, fn, fn)
-    op_dense = MatVecOperator(dense_mat, fn, fn)
-
-    yarr_sparse = sparse_mat.dot(xarr)
-    yarr_dense = dense_mat.dot(xarr)
-
-    # Out-of-place
-    y = op_sparse(x)
-    assert all_almost_equal(y, yarr_sparse)
-
-    y = op_dense(x)
-    assert all_almost_equal(y, yarr_dense)
-
-    # In-place
-    y = fn.element()
-    op_sparse(x, out=y)
-    assert all_almost_equal(y, yarr_sparse)
-
-    y = fn.element()
-    op_dense(x, out=y)
-    assert all_almost_equal(y, yarr_dense)
-
-    # Rectangular case
-    rect_mat = 2 * np.eye(2, 3)
-    r2, r3 = odl.rn(2), odl.rn(3)
-
-    op = MatVecOperator(rect_mat, r3, r2)
-    xarr = np.arange(3, dtype=float)
-    x = r3.element(xarr)
-
-    yarr = rect_mat.dot(xarr)
-
-    # Out-of-place
-    y = op(x)
-    assert all_almost_equal(y, yarr)
-
-    # In-place
-    y = r2.element()
-    op(x, out=y)
-    assert all_almost_equal(y, yarr)
-
-
 # --- Weighting tests --- #
 
 
@@ -1040,7 +837,7 @@ def test_matrix_matrix():
     w_sparse = NumpyFnMatrixWeighting(sparse_mat)
     w_dense = NumpyFnMatrixWeighting(dense_mat)
 
-    assert isinstance(w_sparse.matrix, sp.sparse.spmatrix)
+    assert isinstance(w_sparse.matrix, scipy.sparse.spmatrix)
     assert isinstance(w_dense.matrix, np.ndarray)
 
 
@@ -1122,20 +919,20 @@ def test_matrix_equiv():
     assert not w_dense.equiv(w_different_dense)
 
     # Test shortcuts
-    sparse_eye = sp.sparse.eye(5)
+    sparse_eye = scipy.sparse.eye(5)
     w_eye = NumpyFnMatrixWeighting(sparse_eye)
     w_dense_eye = NumpyFnMatrixWeighting(sparse_eye.todense())
     w_eye_vec = NumpyFnArrayWeighting(np.ones(5))
 
     w_eye_wrong_exp = NumpyFnMatrixWeighting(sparse_eye, exponent=1)
 
-    sparse_smaller_eye = sp.sparse.eye(4)
+    sparse_smaller_eye = scipy.sparse.eye(4)
     w_smaller_eye = NumpyFnMatrixWeighting(sparse_smaller_eye)
 
-    sparse_shifted_eye = sp.sparse.eye(5, k=1)
+    sparse_shifted_eye = scipy.sparse.eye(5, k=1)
     w_shifted_eye = NumpyFnMatrixWeighting(sparse_shifted_eye)
 
-    sparse_almost_eye = sp.sparse.dia_matrix((np.ones(4), [0]), (5, 5))
+    sparse_almost_eye = scipy.sparse.dia_matrix((np.ones(4), [0]), (5, 5))
     w_almost_eye = NumpyFnMatrixWeighting(sparse_almost_eye)
 
     assert w_eye.equiv(w_dense_eye)
@@ -1201,7 +998,7 @@ def test_matrix_norm(fn, exponent):
         true_norm_dense = np.linalg.norm(dense_mat.dot(xarr), ord=exponent)
     else:  # ||x||_{A,p} = ||A^{1/p} x||_p
         # Calculate matrix power
-        eigval, eigvec = sp.linalg.eigh(dense_mat)
+        eigval, eigvec = scipy.linalg.eigh(dense_mat)
         eigval **= 1.0 / exponent
         mat_pow = (eigval * eigvec).dot(eigvec.conj().T)
         true_norm_dense = np.linalg.norm(np.dot(mat_pow, xarr), ord=exponent)
@@ -1249,7 +1046,7 @@ def test_matrix_dist(fn, exponent):
                                          ord=exponent)
     else:  # d(x, y)_{A,p} = ||A^{1/p} (x-y)||_p
         # Calculate matrix power
-        eigval, eigvec = sp.linalg.eigh(dense_mat)
+        eigval, eigvec = scipy.linalg.eigh(dense_mat)
         eigval **= 1.0 / exponent
         mat_pow = (eigval * eigvec).dot(eigvec.conj().T)
         true_dist_dense = np.linalg.norm(np.dot(mat_pow, xarr - yarr),
@@ -1483,8 +1280,8 @@ def test_constant_equals():
     w_other_const = NumpyFnConstWeighting(constant + 1)
     w_other_exp = NumpyFnConstWeighting(constant, exponent=1)
 
-    const_sparse_mat = sp.sparse.dia_matrix(([constant] * n, [0]),
-                                            shape=(n, n))
+    const_sparse_mat = scipy.sparse.dia_matrix(([constant] * n, [0]),
+                                               shape=(n, n))
     const_dense_mat = constant * np.eye(n)
     w_matrix_sp = NumpyFnMatrixWeighting(const_sparse_mat)
     w_matrix_de = NumpyFnMatrixWeighting(const_dense_mat)
@@ -1508,8 +1305,8 @@ def test_constant_equiv():
     w_const = NumpyFnConstWeighting(constant)
     w_const2 = NumpyFnConstWeighting(constant)
 
-    const_sparse_mat = sp.sparse.dia_matrix(([constant] * n, [0]),
-                                            shape=(n, n))
+    const_sparse_mat = scipy.sparse.dia_matrix(([constant] * n, [0]),
+                                               shape=(n, n))
     const_dense_mat = constant * np.eye(n)
     w_matrix_sp = NumpyFnMatrixWeighting(const_sparse_mat)
     w_matrix_de = NumpyFnMatrixWeighting(const_dense_mat)
